Don't relayout app contents when just changing app position
In WMS.resizeTasks we call task.resizeWindows() whenever the
task bounds changes which causes the app to do layout passes.
This isn't needed if we are just changing the position of the
task and not the size. This is currently causing unneeded churn
in the system and which leads to lag with the dragging operation.

Bug: 23901900
Change-Id: I339e31af3b657db6146dc1220bf5eb13e18b7876