Make home stack move like another stack.
Previously the home stack was either positioned at the top or bottom
of the stack list. This type of movement was okay when there was mainly
2 stacks (home and app), but doesn't work in all cases where we have
multiple stacks.
For example:
1. Launch any activity in the freeform stack
2. Press the home button
3. Launch a translucent or dialog activity
In this case you would except the home stack to be seen behind the
translucent activity since that is the stack the user is coming from,
but you see the freeform stack instead since the home stack was moved
to the bottom of all stacks when the translucent activity was launched.
We no longer move the home stack to the bottom when a new
activity/task/stack comes to the front.

Bug: 23626353
Change-Id: Ic506acc8528e63b8e7f999dd88a450bee6fb4552