docs: Removing reference to Apache HTTP client
With Android M, the Apache HTTP client is no longer supported.
Removing references to that client and the clients that depend
on it (DefaultHttpClient and AndroidHttpClient). Replacing, when
appropriate, with references to HttpURLConnection.

Also removing references to the 2011 blog post "Android's HTTP
Clients" (https://goto.google.com/cusll) since that post is largely
about the ex-clients.

Fixed a couple of other small errors while I had the files open
(for example, changed a sample URL in a code snippet to the safe
www.example.com).

bug: 19696844
Change-Id: Iabf780140c37d5fc95c0d45542c425a945c69337