Improve docked & freeform task to front app transition.
Current transition animation assumes that it will go into fullscreen
mode. This makes it aware of the docked stack and makes it translate
from the click position to the final position. It also is aware of the
freeform bounds, so it can translate from the click position to the
freeform position.

Bug: 24913915
Change-Id: Idd42369e3ea86f85de851e2797ca9cef4b3842c2