Fix touching outside of popup not dismissing it.
We only included the exact popup window area for intercepting touches
that would dismiss the popup. We need to intercept the parent area, so
we use the task or stack associated with the window to determine which
touch events to intercept and use as dismiss. This doesn't affect which
touch events will be delivered to the window.

Bug: 25329650

Change-Id: Ie248ec67bb5328fd67b470f556cf206ce888514c