Fix missing dock divider.
When invisible the dock divider would have its bounds adjusted to size
0, 0 and that would be sent to Surface Flinger. When it becomes visible
we immediately adjust the size of the surface, but it's too late for
Surface Flinger and it returns wrong size to the renderer. By not
adjusting the dock divider bounds when it's invisible, Surface Flinger
will always hold good size. The position will be adjusted when the
divider gets visible.

Bug: 25612050

Change-Id: Iab0798d6545ae0f8b46eabfe7e4100079779bd06